05-tests/requirements.txt
pytest
pytest-asyncio
httpx
fastapi
sqlalchemy
python-dotenv
pydantic
psycopg[binary]
PyJWT